feat: 블록에디터 컴포넌트 모듈
